refactor: scssへの切り出しとディレクトリ構成の変更
