+++ frontend/src/app/__tests__/loading.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import Loading from "./loading";
+import Loading from "../loading";
 
 describe("Loading", () => {
   it("renders loading message", () => {
