+++ frontend/src/app/__tests__/not-found.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import NotFound from "./not-found";
+import NotFound from "../not-found";
 
 describe("NotFound", () => {
   it("renders 404 heading and message", () => {
