+++ frontend/src/app/loading.tsx
@@ -1,6 +1,8 @@
+import styles from "./styles/loading.module.scss";
+
 export default function Loading() {
   return (
-    <div className="flex min-h-screen items-center justify-center">
+    <div className={styles.container}>
       <p>読み込み中...</p>
     </div>
   );
