+++ frontend/src/app/not-found.tsx
@@ -1,11 +1,12 @@
 import Link from "next/link";
+import styles from "./styles/not-found.module.scss";
 
 export default function NotFound() {
   return (
-    <div className="flex min-h-screen flex-col items-center justify-center gap-4">
-      <h1 className="text-4xl font-bold">404</h1>
+    <div className={styles.container}>
+      <h1 className={styles.heading}>404</h1>
       <p>ページが見つかりません</p>
-      <Link href="/" className="underline hover:text-primary transition-colors">
+      <Link href="/" className={styles.homeLink}>
         ホームに戻る
       </Link>
     </div>
