+++ frontend/src/app/page.tsx
@@ -5,10 +5,11 @@ import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/com
 import { Input } from "@/components/ui/input";
 import { Separator } from "@/components/ui/separator";
 import { Skeleton } from "@/components/ui/skeleton";
+import styles from "./styles/page.module.scss";
 
 export default function HomePage() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-center p-24">
+    <main className={styles.main}>
       <Card className="w-full max-w-2xl">
         <CardHeader>
           <CardTitle className="text-4xl">Plot Platform</CardTitle>
@@ -17,31 +18,31 @@ export default function HomePage() {
           </CardDescription>
         </CardHeader>
         <CardContent className="space-y-4">
-          <p className="text-muted-foreground">🚧 プロジェクト基盤構築中...</p>
+          <p className={styles.statusText}>🚧 プロジェクト基盤構築中...</p>
 
           <Separator />
 
-          <div className="space-y-2">
-            <h3 className="font-semibold">shadcn/ui コンポーネントテスト</h3>
-            <div className="flex flex-wrap gap-2">
+          <div className={styles.section}>
+            <h3 className={styles.sectionTitle}>shadcn/ui コンポーネントテスト</h3>
+            <div className={styles.buttonGroup}>
               <Button variant="default">Default Button</Button>
               <Button variant="outline">Outline Button</Button>
               <Button variant="ghost">Ghost Button</Button>
             </div>
-            <div className="flex gap-2 items-center">
+            <div className={styles.badgeRow}>
               <Badge>Badge</Badge>
               <Badge variant="secondary">Secondary</Badge>
               <Badge variant="outline">Outline</Badge>
             </div>
             <Input placeholder="Input コンポーネント" />
-            <div className="flex items-center gap-2">
+            <div className={styles.avatarRow}>
               <Avatar>
                 <AvatarImage src="https://github.com/shadcn.png" />
                 <AvatarFallback>CN</AvatarFallback>
               </Avatar>
               <span>Avatar コンポーネント</span>
             </div>
-            <div className="space-y-2">
+            <div className={styles.skeletonGroup}>
               <Skeleton className="h-4 w-full" />
               <Skeleton className="h-4 w-3/4" />
             </div>
@@ -49,7 +50,7 @@ export default function HomePage() {
 
           <Separator />
 
-          <p className="text-sm text-muted-foreground">
+          <p className={styles.footerText}>
             ✅ 全 shadcn/ui コンポーネントが正常にインポートされました
           </p>
         </CardContent>
