+++ frontend/src/app/styles/error.module.scss
@@ -1,0 +1,22 @@
+.container {
+    display: flex;
+    min-height: 100vh;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 1rem;
+}
+
+.heading {
+    font-size: 1.5rem;
+    font-weight: 700;
+}
+
+.retryButton {
+    text-decoration: underline;
+    transition: color 0.2s ease;
+
+    &:hover {
+        color: hsl(var(--primary));
+    }
+}
