+++ frontend/src/app/styles/loading.module.scss
@@ -1,0 +1,6 @@
+.container {
+    display: flex;
+    min-height: 100vh;
+    align-items: center;
+    justify-content: center;
+}
