+++ frontend/src/styles/_mixins.scss
@@ -1,0 +1,51 @@
+// 共通 Mixin
+// docs/frontend/05-styling-guide.md に基づく定義
+@use "variables" as *;
+
+@mixin respond-to($bp) {
+    @if $bp ==sm {
+        @media (min-width: $breakpoint-sm) {
+            @content;
+        }
+    }
+
+    @if $bp ==md {
+        @media (min-width: $breakpoint-md) {
+            @content;
+        }
+    }
+
+    @if $bp ==lg {
+        @media (min-width: $breakpoint-lg) {
+            @content;
+        }
+    }
+
+    @if $bp ==xl {
+        @media (min-width: $breakpoint-xl) {
+            @content;
+        }
+    }
+}
+
+@mixin text-ellipsis($lines: 1) {
+    overflow: hidden;
+    text-overflow: ellipsis;
+
+    @if $lines ==1 {
+        white-space: nowrap;
+    }
+
+    @else {
+        display: -webkit-box;
+        -webkit-line-clamp: $lines;
+        -webkit-box-orient: vertical;
+    }
+}
+
+@mixin focus-ring {
+    &:focus-visible {
+        outline: 2px solid hsl(var(--ring));
+        outline-offset: 2px;
+    }
+}
